Put the layout in the general component
