+++ packages/client/src/pages/errors/Errors.module.scss
