+++ packages/client/src/pages/errors/NotFound.tsx
@@ -1,21 +1,6 @@
 import { FC } from 'react'
-import classes from './NotFound.module.scss'
-import { Link, useNavigate } from 'react-router-dom'
+import { GeneralErrorLayout } from 'pages/errors/GeneralErrorLayout'
 
 export const NotFound: FC = () => {
-  const navigate = useNavigate()
-
-  function goBack() {
-    return navigate(-1)
-  }
-
-  return (
-    <div className={classes.root}>
-      <div className={classes.title}>400</div>
-      <div className={classes.text}>page not found</div>
-      <Link to="#" onClick={goBack} className={classes.link}>
-        Back
-      </Link>
-    </div>
-  )
+  return <GeneralErrorLayout name="404" description="page not found" />
 }
